requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.2
asciitree==0.3.3
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
backcall==0.2.0
basicsr==1.3.4.9
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.3
cycler==0.11.0
dask==2024.3.1
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
distributed==2024.3.1
dominate==2.8.0
easydict==1.10
einops==0.6.1
exceptiongroup==1.1.3
executing==2.0.0
facexlib==0.3.0
fasteners==0.19
filelock==3.12.4
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gast==0.5.4
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grad-cam==1.4.3
grpcio==1.58.0
h5py==3.10.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
image-similarity-measures==0.3.6
imageio==2.31.4
imgaug==0.4.0
importlib-metadata==6.8.0
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==2.4
keras==3.1.1
keras-vggface==0.6
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
libclang==18.1.1
lightning-utilities==0.9.0
llvmlite==0.41.0
lmdb==1.3.0
locket==1.0.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.3.2
mmcv-full==1.7.1
mmdet==2.28.0
model-index==0.1.11
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.15
namex==0.0.7
natsort==8.4.0
network==0.1
networks==0.3.7
networkx==3.1
nima==0.9.1
numba==0.58.0
numcodecs==0.12.1
numpy==1.25.0
oauthlib==3.2.2
onnx==1.16.0
onnx-simplifier==0.4.36
onnxruntime==1.17.3
openai-clip==1.0.1
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.25
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
partd==1.4.1
path==16.10.0
pexpect==4.8.0
phasepack==1.5
pickleshare==0.7.5
Pillow==9.1.1
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.8
ptflops==0.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.16.1
pyiqa==0.1.7
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-msssim==1.0.0
pytorch-wavelets==1.3.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.11.2
seaborn==0.13.2
setuptools==60.2.0
shapely==2.0.1
sigfig==1.3.3
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tb-nightly==2.15.0a20230926
tblib==3.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorboardX==2.5.1
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2023.9.18
timm==0.5.4
tomli==2.0.1
toolz==0.12.1
tornado==6.3.3
tqdm==4.61.2
traitlets==5.11.2
ttach==0.0.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
visdom==0.2.4
wcwidth==0.2.8
websocket-client==1.6.4
Werkzeug==2.3.7
wheel==0.38.4
wrapt==1.16.0
xxhash==3.4.1
yapf==0.40.1
yarl==1.9.2
zarr==2.17.1
zict==3.0.0
zipp==3.17.0
